Add size checking to build script

+++ build.sh
@@ -27,6 +27,13 @@ nasm "${BOOTLOADER_SRC}" -f bin -o "${BOOTLOADER_BIN}" \
      -i "${BOOTLOADER_DIR}" \
      -i './macros/'
 
+# check bootloader size is not larger than intended
+BOOTLOADER_SIZE_CHECK=$(wc -c < "${BOOTLOADER_BIN}")
+if [ ${BOOTLOADER_SIZE_CHECK} -gt ${BOOTLOADER_SIZE} ]; then
+    echo "Bootloader size exceeds ${BOOTLOADER_SIZE}: ${BOOTLOADER_SIZE_CHECK}"
+    exit 1
+fi
+
 # copy to bootable hdd image
 dd if="${BOOTLOADER_BIN}" of="${HDD_FILE}" bs=$BOOTLOADER_SIZE count=1
  
@@ -40,6 +47,13 @@ KERNEL_SIZE=512
 # assemble
 nasm "${KERNEL_SRC}" -f bin -o "${KERNEL_BIN}" \
      -i "${KERNEL_DIR}"
+
+# check kernel size is not larger than intended
+KERNEL_SIZE_CHECK=$(wc -c < "${KERNEL_BIN}")
+if [ ${KERNEL_SIZE_CHECK} -gt ${KERNEL_SIZE} ]; then
+    echo "Kernel size exceeds ${KERNEL_SIZE}: ${KERNEL_SIZE_CHECK}"
+    exit 1
+fi
 
 # copy to bootable hdd image
 dd if="${KERNEL_BIN}" of="${HDD_FILE}" obs=$KERNEL_SIZE oseek=1
